feat: Fade list fading durations now specified in bluetooth command
This lets you program both the list of colours and the speed of each
fading animation between each colour remotely.

Durations are in tenths of a second for now.

An example bytestream input would be (in hex):

23 32 FF 00 00 02 07 00 FF 00 0A 00

23 		// start byte
32 		// ascii '2', meaning this is a new fade list
FF 00 00 	// Red
02		// for 2 tenths of a second
07		// Another colour follows
00 FF 00	// Green
0A		// for 10 tenths of a second, or 1 second
00		// No more colours to follow

This would fade quickly from red to green,
then slowly from green back to red, and repeat.